requirements.txt
google-generativeai
speechrecognition
streamlit
asyncio
edge_tts
pygame
numpy 



